gcc fix

+++ include/glaze/ext/eigen.hpp
@@ -41,7 +41,7 @@ namespace glz
          template <auto Opts>
          static void op(auto& value, is_context auto&& ctx, auto&& it, auto&& end)
          {
-            constexpr uint8_t layout = uint8_t(T::Flags & Eigen::ColMajor) << 6;
+            constexpr uint8_t layout = uint8_t(uint64_t(T::Flags) & uint64_t(Eigen::ColMajor)) << 6;
             if ((uint8_t(*it) & 0b11000000) != layout) {
                ctx.error = error_code::syntax_error;
             }
@@ -60,7 +60,7 @@ namespace glz
          template <auto Opts>
          static void op(auto& value, is_context auto&& ctx, auto&& it, auto&& end)
          {
-            constexpr uint8_t layout = uint8_t(T::Flags & Eigen::ColMajor) << 6;
+            constexpr uint8_t layout = uint8_t(uint64_t(T::Flags) & uint64_t(Eigen::ColMajor)) << 6;
             if ((uint8_t(*it) & 0b11000000) != layout) {
                ctx.error = error_code::syntax_error;
             }
@@ -80,7 +80,7 @@ namespace glz
          static void op(auto&& value, is_context auto&& ctx, auto&&... args) noexcept
          {
             constexpr uint8_t matrix = 0b00'010'000;
-            constexpr uint8_t layout = uint8_t(T::Flags & Eigen::ColMajor) << 6;
+            constexpr uint8_t layout = uint8_t(uint64_t(T::Flags) & uint64_t(Eigen::ColMajor)) << 6;
             constexpr uint8_t tag = tag::additional | matrix | layout;
             dump_type(tag, args...);
             
@@ -99,7 +99,7 @@ namespace glz
          static void op(auto&& value, is_context auto&& ctx, auto&&... args) noexcept
          {
             constexpr uint8_t matrix = 0b00'010'000;
-            constexpr uint8_t layout = uint8_t(T::Flags & Eigen::ColMajor) << 6;
+            constexpr uint8_t layout = uint8_t(uint64_t(T::Flags) & uint64_t(Eigen::ColMajor)) << 6;
             constexpr uint8_t tag = tag::additional | matrix | layout;
             dump_type(tag, args...);
